move loading of addons to separate module to not trigger pkg_resource call
